+++ src/ray.rs
@@ -1,4 +1,4 @@
-use crate::vec::{Point3, Vec3};
+use crate::{vec::{Point3, Vec3}, Float};
 
 pub struct Ray {
     orig: Point3,
@@ -18,7 +18,7 @@ impl Ray {
         self.dir
     }
 
-    pub fn at(&self, t: f64) -> Point3 {
+    pub fn at(&self, t: Float) -> Point3 {
         self.orig + t * self.dir
     }
 }
